Remove traces of previewChoice setting, consolidate preview url and preview routing on a common constant/util
