+++ static-assets/components/cstudio-common/common-api.js
@@ -1565,7 +1565,7 @@ var nodeOpen = false,
       /**
        * given a transfer object, open a preview URL
        */
-      openPreview: function(contentTO, windowId, soundTone, incontextEdit, targetWindowId) {
+      openPreview: function(contentTO, windowId, soundTone, inContextEdit, targetWindowId) {
         if (!targetWindowId) {
           // if no target is supplied assume local call
           // basically mimics behavior before target was implmented
@@ -1574,7 +1574,7 @@ var nodeOpen = false,
 
         var url = this.getPreviewUrl(contentTO);
 
-        if (incontextEdit) {
+        if (inContextEdit) {
           window.location.reload();
         } else {
           // remove server name and port etc
@@ -1585,11 +1585,12 @@ var nodeOpen = false,
               url = '/';
             }
           }
-          if (CrafterCMSNext.system.store.getState().preview.previewChoice[CStudioAuthoringContext.site] === '2') {
-            window.location = `/studio/next/preview#/?page=${url}&site=${CStudioAuthoringContext.site}`;
-          } else {
-            window.location = `/studio/preview#/?page=${url}&site=${CStudioAuthoringContext.site}`;
-          }
+          window.location = CrafterCMSNext.util.system.getSystemLink({
+            systemLinkId: 'preview',
+            authoringBase: CrafterCMSNext.system.store.getState().env.authoringBase,
+            site: CStudioAuthoringContext.site,
+            page: url
+          });
         }
       },
