+++ ui/app/src/utils/system.ts
@@ -1,0 +1,38 @@
+/*
+ * Copyright (C) 2007-2021 Crafter Software Corporation. All Rights Reserved.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as published by
+ * the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import { PREVIEW_BASE_URL } from './constants';
+
+export type SystemLinkId = 'preview' | 'siteTools' | 'siteSearch' | 'siteDashboard';
+
+export function getSystemLink({
+  systemLinkId,
+  authoringBase,
+  site,
+  page = '/'
+}: {
+  systemLinkId: SystemLinkId;
+  authoringBase: string;
+  site: string;
+  page?: string;
+}) {
+  return {
+    preview: `${authoringBase}${PREVIEW_BASE_URL}#/?page=${page}&site=${site}`,
+    siteTools: `${authoringBase}/site-config`,
+    siteSearch: `${authoringBase}/search`,
+    siteDashboard: `${authoringBase}/site-dashboard`
+  }[systemLinkId];
+}
